+++ source/Setup/Database.php
@@ -8,73 +8,51 @@
 namespace OxidEsales\EshopCommunity\Setup;
 
 use Exception;
+use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
+use OxidEsales\EshopCommunity\Internal\Framework\Database\CompatibilityChecker\Bridge\DatabaseCheckerBridgeInterface;
 use OxidEsales\EshopCommunity\Setup\Exception\LanguageParamsException;
+use OxidEsales\Facts\Facts;
 use PDO;
 use PDOException;
 use PDOStatement;
-use \OxidEsales\Facts\Facts;
 
-/**
- * Setup database manager class
- */
 class Database extends Core
 {
-    /**
-     * Connection resource object
-     *
-     * @var PDO
-     */
+    /** @var int */
+    public const ERROR_DB_CONNECT = 1;
+    /** @var int */
+    public const ERROR_COULD_NOT_CREATE_DB = 2;
+    /** @var int */
+    public const ERROR_CODE_DBMS_NOT_COMPATIBLE = 3;
+    /** @var int */
+    public const ERROR_CODE_DBMS_NOT_RECOMMENDED = 4;
+    /** @var int */
+    private const ERROR_OPENING_SQL_FILE = 1;
+    /** @var PDO */
     protected $_oConn = null;
+    /** @var ContainerFactory */
+    private $containerFactory;
 
-    /**
-     * Error while opening sql file
-     *
-     * @var int
-     */
-    const ERROR_OPENING_SQL_FILE = 1;
-
-    /**
-     * Error while opening db connection
-     *
-     * @var int
-     */
-    const ERROR_DB_CONNECT = 1;
-
-    /**
-     * Error while creating db
-     *
-     * @var int
-     */
-    const ERROR_COULD_NOT_CREATE_DB = 2;
-
-    /**
-     * MySQL version does not fir requirements
-     *
-     * @var int
-     */
-    const ERROR_MYSQL_VERSION_DOES_NOT_FIT_REQUIREMENTS = 3;
-
-    /**
-     * MySQL version does not fir recommendations
-     *
-     * @var int
-     */
-    const ERROR_MYSQL_VERSION_DOES_NOT_FIT_RECOMMENDATIONS = 4;
+    /** @param ContainerFactory|null $containerFactory */
+    public function __construct(ContainerFactory $containerFactory = null)
+    {
+        $this->containerFactory = $containerFactory ?? ContainerFactory::getInstance();
+    }
 
     /**
      * Executes sql query. Returns query execution resource object
      *
      * @param string $sQ query to execute
      *
+     * @return PDOStatement|int
      * @throws Exception exception is thrown if error occured during sql execution
      *
-     * @return PDOStatement|int
      */
     public function execSql($sQ)
     {
         try {
             $pdo = $this->getConnection();
-            list($sStatement) = explode(" ", ltrim($sQ));
+            list($sStatement) = explode(' ', ltrim($sQ));
             if (in_array(strtoupper($sStatement), ['SELECT', 'SHOW'])) {
                 $oStatement = $pdo->query($sQ);
             } else {
@@ -84,7 +62,7 @@ class Database extends Core
             return $oStatement;
         } catch (PDOException $e) {
             throw new Exception(
-                $this->getInstance("Language")->getText('ERROR_BAD_SQL') . "( $sQ ): {$e->getMessage()}\n"
+                $this->translate('ERROR_BAD_SQL') . "( $sQ ): {$e->getMessage()}\n"
             );
         }
     }
@@ -99,31 +77,31 @@ class Database extends Core
         $oPdo = $this->getConnection();
         try {
             // testing creation
-            $sQ      = "create or replace view oxviewtest as select 1";
+            $sQ = "create or replace view oxviewtest as select 1";
             $oPdo->exec($sQ);
         } catch (PDOException $e) {
             throw new Exception(
-                $this->getInstance("Language")->getText('ERROR_VIEWS_CANT_CREATE') . " {$e->getMessage()}\n"
+                $this->translate('ERROR_VIEWS_CANT_CREATE') . " {$e->getMessage()}\n"
             );
         }
 
         try {
             // testing data selection
-            $sQ      = "SELECT * FROM oxviewtest";
+            $sQ = "SELECT * FROM oxviewtest";
             $oPdo->query($sQ)->closeCursor();
         } catch (PDOException $e) {
             throw new Exception(
-                $this->getInstance("Language")->getText('ERROR_VIEWS_CANT_SELECT') . " {$e->getMessage()}\n"
+                $this->translate('ERROR_VIEWS_CANT_SELECT') . " {$e->getMessage()}\n"
             );
         }
 
         try {
             // testing view dropping
-            $sQ      = "drop view oxviewtest";
+            $sQ = "drop view oxviewtest";
             $oPdo->exec($sQ);
         } catch (PDOException $e) {
             throw new Exception(
-                $this->getInstance("Language")->getText('ERROR_VIEWS_CANT_DROP') . " {$e->getMessage()}\n"
+                $this->translate('ERROR_VIEWS_CANT_DROP') . " {$e->getMessage()}\n"
             );
         }
     }
@@ -141,7 +119,7 @@ class Database extends Core
             $oSetup = $this->getInstance("Setup");
             // problems with file
             $oSetup->setNextStep($oSetup->getStep('STEP_DB_INFO'));
-            throw new Exception(sprintf($this->getInstance("Language")->getText('ERROR_OPENING_SQL_FILE'), $sFilename), Database::ERROR_OPENING_SQL_FILE);
+            throw new Exception(sprintf($this->translate('ERROR_OPENING_SQL_FILE'), $sFilename), Database::ERROR_OPENING_SQL_FILE);
         }
 
         $sQuery = fread($fp, filesize($sFilename));
@@ -184,78 +162,17 @@ class Database extends Core
     }
 
     /**
-     * Opens database connection and returns connection resource object
-     *
-     * @param array $aParams database connection parameters array
-     *
-     * @throws Exception exception is thrown if connection failed or was unable to select database
-     *
-     * @return object
-     */
-    public function openDatabase($aParams)
-    {
-        $aParams = (is_array($aParams) && count($aParams)) ? $aParams : $this->getInstance("Session")->getSessionParam('aDB');
-        if ($this->_oConn === null) {
-            // ok open DB
-            try {
-                $dsn = sprintf('mysql:host=%s;port=%s', $aParams['dbHost'], $aParams['dbPort']);
-                $this->_oConn = new PDO(
-                    $dsn,
-                    $aParams['dbUser'],
-                    $aParams['dbPwd'],
-                    [PDO::MYSQL_ATTR_INIT_COMMAND => 'SET NAMES utf8']
-                );
-                $this->_oConn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-                $this->_oConn->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
-            } catch (PDOException $e) {
-                /** @var Setup $oSetup */
-                $oSetup = $this->getInstance("Setup");
-                $oSetup->setNextStep($oSetup->getStep('STEP_DB_INFO'));
-                throw new Exception($this->getInstance("Language")->getText('ERROR_DB_CONNECT') . " - " . $e->getMessage(), Database::ERROR_DB_CONNECT, $e);
-            }
-
-            // testing version
-            $oSysReq = getSystemReqCheck();
-            if (0 === $oSysReq->checkMysqlVersion($this->getDatabaseVersion())) {
-                throw new Exception($this->getInstance("Language")->getText('ERROR_MYSQL_VERSION_DOES_NOT_FIT_REQUIREMENTS'), Database::ERROR_MYSQL_VERSION_DOES_NOT_FIT_REQUIREMENTS);
-            }
-
-            $databaseConnectProblem = null;
-            try {
-                $this->_oConn->exec("USE `{$aParams['dbName']}`");
-            } catch (Exception $e) {
-                $databaseConnectException = new Exception($this->getInstance("Language")->getText('ERROR_COULD_NOT_CREATE_DB') . " - " . $e->getMessage(), Database::ERROR_COULD_NOT_CREATE_DB, $e);
-            }
-
-            if ((1 === $oSysReq->checkMysqlVersion($this->getDatabaseVersion())) && !$this->userDecidedIgnoreDBWarning()) {
-                throw new Exception($this->getInstance("Language")->getText('ERROR_MYSQL_VERSION_DOES_NOT_FIT_RECOMMENDATIONS'), Database::ERROR_MYSQL_VERSION_DOES_NOT_FIT_RECOMMENDATIONS);
-            }
-
-            if (is_a($databaseConnectException, 'Exception')) {
-                throw $databaseConnectException;
-            }
-        } else {
-            // Make sure the database is connected
-            $this->connectDb($aParams['dbName']);
-        }
-
-        return $this->_oConn;
-    }
-
-    /**
-     * Connect to database.
-     *
-     * @param string $sDbName
-     *
+     * @param $parameters
+     * @return PDO
      * @throws Exception
      */
-    public function connectDb($sDbName)
+    public function openDatabase($parameters)
     {
-        try {
-            $this->_oConn->exec("USE `$sDbName`");
-        } catch (Exception $e) {
-            throw new Exception($this->getInstance("Language")->getText('ERROR_COULD_NOT_CREATE_DB') . " - " . $e->getMessage(), Database::ERROR_COULD_NOT_CREATE_DB, $e);
-        }
+        $connectionParameters = $this->prepareConnectionParameters($parameters);
+        $this->preparePdoConnection($connectionParameters);
+        $this->checkDatabaseCompatibility();
+        $this->executeUseStatement($connectionParameters['dbName']);
+        return $this->_oConn;
     }
 
     /**
@@ -269,11 +186,11 @@ class Database extends Core
     {
         try {
             $this->execSql("CREATE DATABASE `$sDbName` CHARACTER SET utf8 COLLATE utf8_general_ci;");
-            $this->connectDb($sDbName);
+            $this->executeUseStatement($sDbName);
         } catch (Exception $e) {
             $oSetup = $this->getInstance("Setup");
             $oSetup->setNextStep($oSetup->getStep('STEP_DB_INFO'));
-            throw new Exception(sprintf($this->getInstance("Language")->getText('ERROR_COULD_NOT_CREATE_DB'), $sDbName) . " - " . $e->getMessage(), $e->getCode(), $e);
+            throw new Exception(sprintf($this->translate('ERROR_COULD_NOT_CREATE_DB'), $sDbName) . " - " . $e->getMessage(), $e->getCode(), $e);
         }
     }
 
@@ -435,7 +352,7 @@ class Database extends Core
      * Updates default admin user login name and password
      *
      * @param string $sLoginName admin user login name
-     * @param string $sPassword  admin user login password
+     * @param string $sPassword admin user login password
      */
     public function writeAdminLoginData($sLoginName, $sPassword)
     {
@@ -453,10 +370,10 @@ class Database extends Core
     /**
      * Adds config value if shop info should be set.
      *
-     * @param Utilities $utilities  Setup utilities
-     * @param string    $baseShopId Shop id
-     * @param array     $parameters Parameters
-     * @param Session   $session    Setup session manager
+     * @param Utilities $utilities Setup utilities
+     * @param string $baseShopId Shop id
+     * @param array $parameters Parameters
+     * @param Session $session Setup session manager
      */
     protected function addConfigValueIfShopInfoShouldBeSent($utilities, $baseShopId, $parameters, $session)
     {
@@ -468,5 +385,120 @@ class Database extends Core
             "insert into oxconfig (oxid, oxshopid, oxvarname, oxvartype, oxvarvalue)
                              values('$sID', '$baseShopId', 'blSendShopDataToOxid', 'bool', '$blSendShopDataToOxid')"
         );
+    }
+
+    /**
+     * @param $parameters
+     * @return array
+     */
+    private function prepareConnectionParameters($parameters): array
+    {
+        return (is_array($parameters) && !empty($parameters)) ?
+            $parameters :
+            $this->getInstance('Session')->getSessionParam('aDB');
+    }
+
+    /**
+     * @param array $connectionParameters
+     * @throws Exception
+     */
+    private function preparePdoConnection(array $connectionParameters): void
+    {
+        if ($this->_oConn !== null) {
+            return;
+        }
+        try {
+            $this->createPdoConnection($connectionParameters);
+        } catch (PDOException $e) {
+            $this->resetSetupStep();
+            throw new Exception(
+                $this->translate('ERROR_DB_CONNECT') . ' - ' . $e->getMessage(),
+                Database::ERROR_DB_CONNECT,
+                $e
+            );
+        }
+    }
+
+    /** @param array $parameters */
+    private function createPdoConnection(array $parameters): void
+    {
+        $dsn = sprintf('mysql:host=%s;port=%s', $parameters['dbHost'], $parameters['dbPort']);
+        $this->_oConn = new PDO(
+            $dsn,
+            $parameters['dbUser'],
+            $parameters['dbPwd'],
+            [PDO::MYSQL_ATTR_INIT_COMMAND => 'SET NAMES utf8']
+        );
+        $this->_oConn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+        $this->_oConn->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
+    }
+
+    private function resetSetupStep(): void
+    {
+        $setup = $this->getInstance('Setup');
+        $setup->setNextStep($setup->getStep('STEP_DB_INFO'));
+    }
+
+    /**
+     * @param string $name
+     * @throws Exception
+     */
+    private function executeUseStatement(string $name): void
+    {
+        try {
+            $this->_oConn->exec("USE `$name`");
+        } catch (Exception $e) {
+            throw new Exception(
+                $this->translate('ERROR_COULD_NOT_CREATE_DB') . ' - ' . $e->getMessage(),
+                Database::ERROR_COULD_NOT_CREATE_DB,
+                $e
+            );
+        }
+    }
+
+    /** @throws Exception */
+    private function checkDatabaseCompatibility(): void
+    {
+        $databaseChecker = $this->getDatabaseCheckerBridge();
+        if (!$databaseChecker->isDatabaseCompatible()) {
+            throw new Exception(
+                $this->translate('ERROR_DBMS_VERSION_DOES_NOT_FIT_REQUIREMENTS'),
+                Database::ERROR_CODE_DBMS_NOT_COMPATIBLE
+            );
+        }
+        $notices = $databaseChecker->getCompatibilityNotices();
+        if ($notices && !$this->userDecidedIgnoreDBWarning()) {
+            throw new Exception(
+                $this->getCompatibilityNoticesMessage($notices),
+                Database::ERROR_CODE_DBMS_NOT_RECOMMENDED
+            );
+        }
+    }
+
+    /** @return DatabaseCheckerBridgeInterface */
+    private function getDatabaseCheckerBridge(): DatabaseCheckerBridgeInterface
+    {
+        return $this->containerFactory->getContainer()->get(DatabaseCheckerBridgeInterface::class);
+    }
+
+    /**
+     * @param array $notices
+     * @return string
+     */
+    private function getCompatibilityNoticesMessage(array $notices): string
+    {
+        array_walk($notices, function (&$notice) {
+            $notice = $this->translate($notice);
+        });
+        return implode("\n", $notices);
+    }
+
+    /**
+     * @param string $message
+     * @return string
+     */
+    private function translate(string $message): string
+    {
+        return (string)$this->getInstance('Language')->getText($message);
     }
 }