+++ source/Setup/En/lang.php
@@ -50,8 +50,8 @@ $aLang = [
 'ERROR_PASSWORD_TOO_SHORT'                      => 'Password is too short!',
 'ERROR_PASSWORDS_DO_NOT_MATCH'                  => 'Passwords do not match!',
 'ERROR_USER_NAME_DOES_NOT_MATCH_PATTERN'        => 'Please enter a valid e-mail address!',
-'ERROR_MYSQL_VERSION_DOES_NOT_FIT_REQUIREMENTS' => 'The installed MySQL version does not fit system requirements!',
-'ERROR_MYSQL_VERSION_DOES_NOT_FIT_RECOMMENDATIONS' => 'WARNING: A bug in MySQL 5.6 may lead to problems in OXID eShop Enterprise Edition. Hence, we do not recommend MySQL 5.6. Please see <a href="https://www.oxid-esales.com/de/support-services/dokumentation-und-hilfe/oxid-eshop/installation/oxid-eshop-neu-installieren/server-und-systemvoraussetzungen/systemvoraussetzungen-ee.html">system requirements for OXID eShop Enterprise Edition</a>.',
+'ERROR_DBMS_VERSION_DOES_NOT_FIT_REQUIREMENTS' => 'The version of the installed database management system (DBMS) does not fit system requirements!',
+'ERROR_MYSQL_56_NOT_RECOMMENDED' => 'WARNING: A bug in MySQL 5.6 may lead to problems in OXID eShop Enterprise Edition. Hence, we do not recommend MySQL 5.6. Please see <a href="https://www.oxid-esales.com/de/support-services/dokumentation-und-hilfe/oxid-eshop/installation/oxid-eshop-neu-installieren/server-und-systemvoraussetzungen/systemvoraussetzungen-ee.html">system requirements for OXID eShop Enterprise Edition</a>.',
 
 'ERROR_VIEWS_CANT_CREATE'                       => 'ERROR: Can\'t create views. Please check your database user privileges.',
 'ERROR_VIEWS_CANT_SELECT'                       => 'ERROR: Can\'t select from views. Please check your database user privileges.',