+++ tests/Unit/Core/SystemRequirementsTest.php
@@ -9,12 +9,13 @@ namespace OxidEsales\EshopCommunity\Tests\Unit\Core;
 
 use OxidEsales\Eshop\Core\SystemRequirements;
 use Psr\Container\ContainerInterface;
+use PHPUnit\Framework\MockObject\MockObject as Mock;
 
 class SystemRequirementsTest extends \OxidTestCase
 {
     public function testGetBytes()
     {
-        $systemRequirements = new \OxidEsales\Eshop\Core\SystemRequirements();
+        $systemRequirements = new SystemRequirements();
 
         $this->assertEquals(33554432, $systemRequirements->UNITgetBytes('32M'));
         $this->assertEquals(32768, $systemRequirements->UNITgetBytes('32K'));
@@ -23,7 +24,7 @@ class SystemRequirementsTest extends \OxidTestCase
 
     public function testGetRequiredModules()
     {
-        $systemRequirements = new \OxidEsales\Eshop\Core\SystemRequirements();
+        $systemRequirements = new SystemRequirements();
 
         $requiredModules = $systemRequirements->getRequiredModules();
         $this->assertTrue(is_array($requiredModules));
@@ -34,8 +35,8 @@ class SystemRequirementsTest extends \OxidTestCase
 
     public function testGetModuleInfo()
     {
-        /** @var SystemRequirements|PHPUnit\Framework\MockObject\MockObject $systemRequirementsMock */
-        $systemRequirementsMock = $this->getMock(\OxidEsales\Eshop\Core\SystemRequirements::class, array('checkMbString', 'checkModRewrite'));
+        /** @var SystemRequirements|Mock $systemRequirementsMock */
+        $systemRequirementsMock = $this->getMock(SystemRequirements::class, array('checkMbString', 'checkModRewrite'));
 
         $systemRequirementsMock->expects($this->once())->method('checkMbString');
         $systemRequirementsMock->expects($this->never())->method('checkModRewrite');
@@ -83,7 +84,7 @@ class SystemRequirementsTest extends \OxidTestCase
      */
     public function testCheckMysqlVersion($version, $expectedResult)
     {
-        $systemRequirements = new \OxidEsales\Eshop\Core\SystemRequirements();
+        $systemRequirements = new SystemRequirements();
 
         $this->assertEquals($expectedResult, $systemRequirements->checkMysqlVersion($version));
     }
@@ -136,7 +137,7 @@ class SystemRequirementsTest extends \OxidTestCase
 
     public function testCheckCollation()
     {
-        $systemRequirements = new \OxidEsales\Eshop\Core\SystemRequirements();
+        $systemRequirements = new SystemRequirements();
 
         $collations = $systemRequirements->checkCollation();
 
@@ -145,8 +146,8 @@ class SystemRequirementsTest extends \OxidTestCase
 
     public function testGetSysReqStatus()
     {
-        /** @var SystemRequirements|PHPUnit\Framework\MockObject\MockObject $systemRequirementsMock */
-        $systemRequirementsMock = $this->getMock(\OxidEsales\Eshop\Core\SystemRequirements::class, array('getSystemInfo'));
+        /** @var SystemRequirements|Mock $systemRequirementsMock */
+        $systemRequirementsMock = $this->getMock(SystemRequirements::class, array('getSystemInfo'));
         $systemRequirementsMock->expects($this->once())->method('getSystemInfo');
 
         $this->assertTrue($systemRequirementsMock->getSysReqStatus());
@@ -160,7 +161,7 @@ class SystemRequirementsTest extends \OxidTestCase
     public function testGetReqInfoUrl()
     {
         $sUrl = "https://oxidforge.org/en/system-requirements";
-        $systemRequirements = new \OxidEsales\Eshop\Core\SystemRequirements();
+        $systemRequirements = new SystemRequirements();
 
         $this->assertEquals($sUrl . "#PHP_version_at_least_7.0", $systemRequirements->getReqInfoUrl("php_version"));
         $this->assertEquals($sUrl, $systemRequirements->getReqInfoUrl("none"));
@@ -175,7 +176,7 @@ class SystemRequirementsTest extends \OxidTestCase
     public function testGetShopHostInfoFromConfig()
     {
         $this->getConfig()->setConfigParam('sShopURL', 'http://www.testshopurl.lt/testsubdir1/insideit2/');
-        $systemRequirements = new \OxidEsales\Eshop\Core\SystemRequirements();
+        $systemRequirements = new SystemRequirements();
         $this->assertEquals(
             array(
                 'host' => 'www.testshopurl.lt',
@@ -225,7 +226,7 @@ class SystemRequirementsTest extends \OxidTestCase
     public function testGetShopSSLHostInfoFromConfig()
     {
         $this->getConfig()->setConfigParam('sSSLShopURL', 'http://www.testshopurl.lt/testsubdir1/insideit2/');
-        $systemRequirements = new \OxidEsales\Eshop\Core\SystemRequirements();
+        $systemRequirements = new SystemRequirements();
         $this->assertEquals(
             array(
                 'host' => 'www.testshopurl.lt',
@@ -279,7 +280,7 @@ class SystemRequirementsTest extends \OxidTestCase
         $_SERVER['SERVER_PORT'] = null;
         $_SERVER['HTTP_HOST'] = 'www.testshopurl.lt';
 
-        $systemRequirements = new \OxidEsales\Eshop\Core\SystemRequirements();
+        $systemRequirements = new SystemRequirements();
         $this->assertEquals(
             array(
                 'host' => 'www.testshopurl.lt',
@@ -405,8 +406,8 @@ class SystemRequirementsTest extends \OxidTestCase
             'OXMODULE'    => '_OXMODULE_',
         );
 
-        /** @var SystemRequirements|PHPUnit\Framework\MockObject\MockObject $systemRequirementsMock */
-        $systemRequirementsMock = $this->getMock(\OxidEsales\Eshop\Core\SystemRequirements::class, array('_checkTemplateBlock', 'fetchBlockRecords'));
+        /** @var SystemRequirements|Mock $systemRequirementsMock */
+        $systemRequirementsMock = $this->getMock(SystemRequirements::class, array('_checkTemplateBlock', 'fetchBlockRecords'));
         $systemRequirementsMock->expects($this->exactly(1))->method('_checkTemplateBlock')
             ->with($this->equalTo("_OXTEMPLATE_"), $this->equalTo("_OXBLOCKNAME_"))
             ->will($this->returnValue(false));
@@ -441,8 +442,8 @@ class SystemRequirementsTest extends \OxidTestCase
             'OXMODULE'    => '_OXMODULE_',
         );
 
-        /** @var SystemRequirements|PHPUnit\Framework\MockObject\MockObject $systemRequirementsMock */
-        $systemRequirementsMock = $this->getMock(\OxidEsales\Eshop\Core\SystemRequirements::class, array('_checkTemplateBlock', 'fetchBlockRecords'));
+        /** @var SystemRequirements|Mock $systemRequirementsMock */
+        $systemRequirementsMock = $this->getMock(SystemRequirements::class, array('_checkTemplateBlock', 'fetchBlockRecords'));
         $systemRequirementsMock->expects($this->exactly(1))->method('_checkTemplateBlock')
             ->with($this->equalTo("_OXTEMPLATE_"), $this->equalTo("_OXBLOCKNAME_"))
             ->will($this->returnValue(true));
@@ -485,8 +486,8 @@ class SystemRequirementsTest extends \OxidTestCase
      */
     public function testCheckPhpVersion($sVersion, $iResult)
     {
-        /** @var SystemRequirements|PHPUnit\Framework\MockObject\MockObject $systemRequirementsMock */
-        $systemRequirementsMock = $this->getMock(\OxidEsales\Eshop\Core\SystemRequirements::class, array('getPhpVersion'));
+        /** @var SystemRequirements|Mock $systemRequirementsMock */
+        $systemRequirementsMock = $this->getMock(SystemRequirements::class, array('getPhpVersion'));
         $systemRequirementsMock->expects($this->once())->method('getPhpVersion')->will($this->returnValue($sVersion));
 
         $this->assertSame($iResult, $systemRequirementsMock->checkPhpVersion());
@@ -529,7 +530,7 @@ class SystemRequirementsTest extends \OxidTestCase
             $this->markTestSkipped('This test is for Community and Professional editions only.');
         }
 
-        $systemRequirements = new \OxidEsales\Eshop\Core\SystemRequirements();
+        $systemRequirements = new SystemRequirements();
 
         $this->assertEquals($expectedResult, $systemRequirements->checkMemoryLimit($memoryLimit));
     }